Refactor client
Extract multistatus response
