+++ src/dav/client.ts
@@ -4,28 +4,11 @@ import axios, {
     AxiosRequestConfig,
 } from 'axios'
 import { FileProps } from './fileProps'
-import { DOMParser } from 'xmldom'
 import { Tag } from './tag'
-interface MultiStatusResponse {
-    href: string | null
-    propStat: PropertyStatus[]
-}
-
-interface PropertyStatus {
-    status: string
-    properties: object
-}
-
-type ResolverFunction = (namespace: string) => string | undefined
+import { MultiStatusResponse } from './multiStatusResponse'
 
 export class Client {
     constructor(readonly connection: AxiosInstance) {}
-    private xmlNamespaces: object = {
-        'DAV:': 'd',
-        'http://owncloud.org/ns': 'oc',
-        'http://nextcloud.org/ns': 'nc',
-        'http://open-collaboration-services.org/ns': 'ocs',
-    }
 
     addTag = async (fileId: string, tag: Tag) =>
         this.connection({
@@ -195,118 +178,7 @@ export class Client {
     }
 
     private _parseMultiStatus = (doc: string): MultiStatusResponse[] => {
-        const result: MultiStatusResponse[] = []
-        const xmlNamespaces: object = this.xmlNamespaces
-        const resolver: ResolverFunction = (namespace: string) => {
-            let ii: string
-            for (ii in xmlNamespaces) {
-                if (xmlNamespaces[ii] === namespace) {
-                    return ii
-                }
-            }
-            return undefined
-        }
-
-        const responses = this._getElementsByTagName(
-            doc,
-            'd:response',
-            resolver,
-        )
-        for (let i = 0; i < responses.length; i++) {
-            const responseNode: any = responses[i]
-            const response: MultiStatusResponse = {
-                href: null,
-                propStat: [],
-            }
-
-            const hrefNode: any = this._getElementsByTagName(
-                responseNode,
-                'd:href',
-                resolver,
-            )[0]
-
-            response.href = hrefNode.textContent || hrefNode.text
-
-            const propStatNodes = this._getElementsByTagName(
-                responseNode,
-                'd:propstat',
-                resolver,
-            )
-
-            for (let j = 0; j < propStatNodes.length; j++) {
-                const propStatNode: any = propStatNodes[j]
-                const statusNode: any = this._getElementsByTagName(
-                    propStatNode,
-                    'd:status',
-                    resolver,
-                )[0]
-
-                const propStat: PropertyStatus = {
-                    status: statusNode.textContent || statusNode.text,
-                    properties: {},
-                }
-
-                const propNode: any = this._getElementsByTagName(
-                    propStatNode,
-                    'd:prop',
-                    resolver,
-                )[0]
-                if (!propNode) {
-                    continue
-                }
-                for (let k = 0; k < propNode.childNodes.length; k++) {
-                    const prop: any = propNode.childNodes[k]
-                    const value: any = this._parsePropNode(prop)
-                    const namespace: string =
-                        this.xmlNamespaces[prop.namespaceURI] ||
-                        prop.namespaceURI
-                    propStat.properties[
-                        `${namespace}:${prop.localName || prop.baseName}`
-                    ] = value
-                }
-                response.propStat.push(propStat)
-            }
-
-            result.push(response)
-        }
-
-        return result
-    }
-
-    private _parsePropNode = (e: any): string => {
-        let t: any[] | null = null
-        if (e.childNodes && e.childNodes.length > 0) {
-            const n: any[] = []
-            for (let r = 0; r < e.childNodes.length; r++) {
-                const i: any = e.childNodes[r]
-                if (1 === i.nodeType) {
-                    n.push(i)
-                }
-            }
-            if (n.length) {
-                t = n
-            }
-        }
-        return t || e.textContent || e.text || ''
-    }
-
-    private _getElementsByTagName = (
-        node: Document | string,
-        name: string,
-        resolver: ResolverFunction,
-    ): HTMLCollectionOf<Element> => {
-        const parts: string[] = name.split(':')
-        const tagName: string = parts[1]
-        // @Sergey what to do here? namespace could be undefined, I put in a naive fix..
-        const namespace: string = resolver(parts[0]) || ''
-        if (typeof node === 'string') {
-            const parser: DOMParser = new DOMParser()
-            node = parser.parseFromString(node, 'text/xml')
-        }
-        if (node.getElementsByTagNameNS) {
-            return node.getElementsByTagNameNS(namespace, tagName)
-        }
-        return node.getElementsByTagName(name)
+        return MultiStatusResponse.fromString(doc);
     }
 
     static create = (config: AxiosRequestConfig): Client => {
